# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API created to pracite golang skills",
        "title": "Simple CRUD API",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
...
        "/student/add/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Take create student request and creates an student",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "parameters": [
                    {
                        "description": "Enter a student data",
                        "name": "Student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/StudentSwagger"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/student/del/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Take Delete student request and deletes an student",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter a student id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/student/list/{id}": {
            "get": {
                "description": "Show data of specific Student",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter a student id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/student/upd/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Take update student request and updates an student data",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter a student id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Enter a student data",
                        "name": "Student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/StudentSwagger"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "StudentSwagger": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 5
                },
                "firstname": {
                    "description": "Student object swaggger",
                    "type": "string",
                    "example": "Student1"
                },
                "lastname": {
                    "type": "string",
                    "example": "Lastname1"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "In this version of swagger you must write \"Bearer \" prefiks before your JWT token \"Bearer [your_token]\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}